First draft of script
The phase 1.1 plan will be to simply get the script running at all, on
a small subset of the documents.

Next up, we'll need to figure out network file access/handling large
volumes of data (the available TDM corpus exceeds the space on this
machine), and visualization/interface. And then for phase 2 we'll need
to look at evaluating training and tokenization, and more advanced
visualization possibilities, and pretrained embeddings, and all sorts
of fun stuff.

But for now, let's get anything working at all.