Add instructions for efficiently grabbing lots of files
Each thesis has a set of several files (original txt, processed txt,
xml, pdf, etc.), located in a thesis-specific directory. We only need
the processed txt and the xml, and we don't need the directory
structure.

This is easy with enough Unix magic. Figuring out the Unix magic is
hard. But now it is documented!