First pass at writing file for d3
We're going to be visualizing the thesis data with d3; this requires
writing a JSON file suitable for consumption by d3.

We didn't originally grab the .xml files, so we don't have the metadata
that d3 would like (e.g. titles) for human-readable display, so this
script will need to be updated.