Add the ability to write networks around a word
The question I'm investigating here is what do concept clusters look
like. In Aero/Astro, 'oxygen' is part of a local network that contains
words like 'hydrogen', 'nitrogen', 'propellant', and 'hypergolic' -
this looks like a concept cluster that roughly means 'rocket fuel'.

That's awesome, but it also seems that 'oxygen' would mean different
things to other disciplines! For example, in chemistry, it likely means
something closer to 'element'.

We're going to add the ability to write json of the network around a
given word, so that we'll be able to visualize it.